% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anubix.R
\name{anubix}
\alias{anubix}
\title{Perform ANUBIX for the provided gene sets}
\usage{
anubix(network, links_matrix, genesets, pathways, cores = 2, cutoff = 0.8,
sampling = 2000, network_type = "weighted")
}
\arguments{
\item{network}{A data.frame. Two columns if the network has no weights, where column 1 and column 2 are genes. Each row means a link between genes. If the network is weighted, then the third column contains the weights of the links.}

\item{links_matrix}{A numeric matrix. Stores the links each gene has to each pathway. The number of rows equals the number of genes in the network and the number of columns equals the number of pathways under study.}

\item{genesets}{A data.frame of two columns. Column 1 contains the genes and column 2 the experiment where they belong.}

\item{pathways}{A data.frame of two columns. Column 1 contains the genes and column 2 the pathway where they belong.}

\item{cores}{A numeric value. Number of cores used by the algorithm. Default is \strong{2}.}

\item{cutoff}{A numeric value. Link confidence threshold of the weights between genes. Default is \strong{0.8}.}

\item{sampling}{A numeric value. Number of random samplings required to construct the null distribution. Default is \strong{2000}.}

\item{network_type}{Either "weighted" or "unweighted". Default is \strong{weighted}.}
}
\value{
A data frame with the following columns:
\itemize{
\item geneset   - Gene set under study.
\item pathway   - Pathway under study.
\item obv_links - Observed number of links between the gene set and the pathway.
\item exp_mean  - Expected number of links between the gene set and the pathway.
\item overlap   - Number of genes shared by the gene set and the pathway.
\item p-value   - p-value of the test.
\item q-value   - Corrected p-value using Benjamini-Hochberg.
\item FWER      - Corrected p-value using Bonferroni correction.
}
}
\description{
Computes ANUBIX, an accurate test for network enrichment analysis between query sets and pathway sets. Instead of normal random sampling it does constrained random sampling, taking the degree of the nodes into account.
}
\note{
anubix_links() needs to be run beforehand
}
\examples{
\dontrun{
 anubix(
   network = example_anubix$network,
   links_matrix = example_anubix$links_genes,
   genesets = example_anubix$gene_set,
   pathways = example_anubix$pathway_set,
   cores = 2,
   cutoff = 0.8,
   sampling = 2000,
   network_type = "weighted"
 )
}
}
\seealso{
\code{\link{anubix_links}}, \code{\link{anubix_transitivity}}, \code{\link{anubix_clustering}}, \code{\link{example_anubix}}
}
