% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anubix_constrained_website.R
\name{anubix_constrained_website}
\alias{anubix_constrained_website}
\title{Perform ANUBIX for the provided gene sets}
\usage{
anubix_constrained_website(
  links_matrix,
  genesets,
  pathways,
  cores = 2,
  sampling = 2000,
  callback = NULL
)
}
\arguments{
\item{links_matrix}{A numeric matrix. Links_matrix stores the links each gene has to each pathway. The number of rows equals to the number of genes in the network and the number of columns equals to the number of pathways under study.}

\item{genesets}{A data.frame of two columns. Column 1, are the genes and column 2 the experiment where they belong to.}

\item{pathways}{A data.frame of two columns. Column 1 are the genes and second column the pathway where they belong to.}

\item{cores}{A numeric value. Cores is defined as the number of cores used by the algorithm. The default value is \strong{2}}

\item{sampling}{A numeric value. Sampling is defined as the number of random samplings required to construct the null distribution. The default value is \strong{2000}}

\item{new_deg_list}{A list with the genes that fall into certain degree}

\item{map_pos_deg_list}{A data frame which maps the position of a degree value in the new_deg_list list}
}
\value{
A data frame with the following columns:
\itemize{
\item geneset   -   Gene set under study.
\item pathway   -   Pathway under study.
\item obv_links -   Observed number of links between the gene set and the pathway.
\item exp_mean  -   Expected number of links between the gene set and the pathway.
\item overlap   -   Number of genes shared by the gene set and the pathway
\item p-value   -   p-value of the test.
\item q-value   -   Corrected p-value using Benjamini-Hochberg.
\item FWER      -   Corrected p-value using Bonferroni correction.
}
}
\description{
Computes ANUBIX, an accurate test for network enrichment analysis between query sets and pathway sets. Instead of normal random sampling it does constrained random sampling, taking the degree of the nodes into account.
}
\seealso{
\code{\link{anubix_links}},\code{\link{example_anubix}},\code{\link{anubix_clustering}},\code{\link{anubix}}
}
