% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anubix_links.R
\name{anubix_links}
\alias{anubix_links}
\title{Create the Link Matrix Between Genes and Pathways}
\usage{
anubix_links(network, cores = 2, pathways, cutoff = 0.8, network_type = "weighted")
}
\arguments{
\item{network}{A data.frame representing the gene network.
\itemize{
\item If unweighted: Two columns (gene1, gene2), where each row represents an edge.
\item If weighted: Three columns, with the third column containing edge weights.
}}

\item{pathways}{A data.frame with two columns:
\itemize{
\item Column 1: gene names.
\item Column 2: corresponding pathway names.
}}

\item{cutoff}{Numeric. Threshold to filter edges in a weighted network by confidence. Default is \strong{0.8}.}

\item{network_type}{Character. Either \code{"weighted"} or \code{"unweighted"}. Default is \strong{"weighted"}.}

\item{cores}{Integer. Number of CPU cores to use for parallel computation. Default is 2.}
}
\value{
A numeric matrix where:
\itemize{
\item Rows correspond to genes in the network.
\item Columns correspond to pathways.
\item Entries represent the number (or sum of weights) of links from each gene to each pathway.
}
}
\description{
Computes the number or strength of connections that each gene in a network has to each pathway.
This matrix can be used for downstream network enrichment analysis.
}
\examples{
\dontrun{
# Example run on provided example data
anubix_links(
  network = example_anubix$network,
  pathways = example_anubix$pathway_set,
  cutoff = 0.8,
  network_type = "weighted"
)
}
}
\seealso{
\code{\link{anubix}},\code{\link{anubix_transitivity}}, \code{\link{example_anubix}}, \code{\link{anubix_clustering}}
}
