% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anubix_links.R
\name{anubix_links}
\alias{anubix_links}
\title{Creates the link_matrix data}
\usage{
anubix_links(network,cores = 2,pathways,cutoff = 0.75,network_type = "weighted")
}
\arguments{
\item{network}{A data.frame. Two columns if the network has no weights. Where column 1 and column 2 are genes. Each row means a link between genes. If the network is weighted, then the third column are the weights of the links.}

\item{pathways}{A data.frame of two columns. Column 1 are the genes and second column the pathway where they belong to.}

\item{cutoff}{A numeric value. Cutoff is defined as the link confidence threshold of the weights between genes. The default value is \strong{0.75}.}

\item{network_type}{Either "weighted" or "unweighted". The default value is \strong{weighted}.}
}
\value{
A matrix with the following dimensions:
\itemize{
\item nrows   -   Total number of genes in the studied network.
\item ncols   -   Total number of pathways under study.
}
}
\description{
Computation of all the links that each gene in the network has to each of the pathways.
}
\examples{

# Example with a tiny network:
\dontrun{
anubix_links(example_anubix$network,example_anubix$pathway_set,cutoff = 0.75, "weighted")
}
}
\seealso{
\code{\link{anubix}},\code{\link{example_anubix}},\code{\link{anubix_clustering}}
}
