% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{example_anubix}
\alias{example_anubix}
\title{Example Data Provided with the Package}
\format{
An object of class \code{list} of length 4.
}
\source{
\itemize{
\item \strong{pathway_set: } Kanehisa, M., and Goto, S. (2002). KEGG: Kyoto Encyclopedia of Genes and Genomes. \emph{Nucleic Acids Research}, 28(1), 27–30.
\item \strong{network: } Persson, E., Castresana-Aguirre, M., Buzzao, D., Guala, D., & Sonnhammer, E. L. L. (2021). FunCoup 5: functional association networks in all domains of life, supporting directed links and tissue-specificity. \emph{Journal of Molecular Biology}, 433(11), 166835. \doi{10.1016/j.jmb.2021.166835}
}
}
\usage{
example_anubix
}
\description{
A list containing example data for running ANUBIX, including a precomputed gene-pathway link matrix,
pathway definitions, a gene set, and a network subset.

The list contains:
\itemize{
\item \strong{links_genes} – A gene-pathway link matrix computed using a confidence cutoff of 0.8 in the functional association network FunCoup 5.
\item \strong{pathway_set} – A data.frame of 288 KEGG pathways for \emph{Homo sapiens}.
\item \strong{gene_set} – An example gene set consisting of 42 genes.
\item \strong{network} – A subset of the FunCoup 5 network for \emph{Homo sapiens}, containing 2000 links and 1788 genes.
}
}
\examples{
\dontrun{
data(example_anubix)

# We provide the precomputed link matrix (FunCoup 5, cutoff = 0.8) to run ANUBIX directly.

# Alternatively, compute the link matrix from the example network:
anubix_links(
  network = example_anubix$network,
  pathways = example_anubix$pathway_set,
  cutoff = 0.75,
  network_type = "weighted"
)

# Run ANUBIX using the example data:
 anubix(
   network = example_anubix$network,
   links_matrix = example_anubix$links_genes,
   genesets = example_anubix$gene_set,
   pathways = example_anubix$pathway_set,
   cores = 2,
   cutoff = 0.8,
   sampling = 2000
 )
}
}
\references{
\itemize{
\item Kanehisa, M., and Goto, S. (2002). KEGG: Kyoto Encyclopedia of Genes and Genomes. \emph{Nucleic Acids Research}, 28(1), 27–30.
\item Persson, E., Castresana-Aguirre, M., Buzzao, D., Guala, D., & Sonnhammer, E. L. L. (2021). FunCoup 5: functional association networks in all domains of life, supporting directed links and tissue-specificity. \emph{Journal of Molecular Biology}, 433(11), 166835. \doi{10.1016/j.jmb.2021.166835}
}
}
\seealso{
\code{\link{anubix}},\code{\link{anubix_transitivity}}, \code{\link{anubix_links}}, \code{\link{anubix_clustering}}
}
\keyword{datasets}
